+++ java/KafkaConnectors/cloudformation/msf-msk-iam-auth.yaml
