+++ java/Windowing/cloudformation/msf-msk-iam-auth-windowing.yaml
@@ -8,37 +8,26 @@ Parameters:
     Default: FLINK-1_15
   CodeBucketArn:
     Type: String
-    Default: "arn:aws:s3:::aksh-code-binaries"
   CodeKey:
     Type: String
     Default: flink/amazon-msf-slidingwindow-app-1.0.jar
-  KafkaBootstrapserver:
-    Type: String
-    Default: boot-z6eo0mfk.c1.kafka-serverless.ap-south-1.amazonaws.com:9098
-  Clustername:
-    Type: String
-    Default: demo-cluster-1
   SecurityGroup:
     Type: AWS::EC2::SecurityGroup::Id
-    Default: sg-28d5f054
   SubnetOne:
     Type: AWS::EC2::Subnet::Id
-    Default: subnet-08710af059f886114
   SubnetTwo:
     Type: AWS::EC2::Subnet::Id
-    Default: subnet-7d90f906
   SubnetThree:
     Type: AWS::EC2::Subnet::Id
-    Default: subnet-02e1e451e78007768
   Parallelism:
     Type: String
     Default: 1
+  KafkaBootstrapserver:
+    Type: String
   SourceKafkaTopic:
     Type: String
-    Default: windowing-source
   SinkKafkaTopic:
     Type: String
-    Default: windowing-sink
 
 Resources:
   FlinkApplication:
@@ -86,16 +75,19 @@ Resources:
             AutoScalingEnabled: true
             Parallelism: !Ref Parallelism
             ParallelismPerKPU: 1
+
   FlinkApplicationLogGroup:
     Type: AWS::Logs::LogGroup
     Properties:
       RetentionInDays: 7
-      LogGroupName: !Sub "/aws/kinesis-analytics/${ApplicationName}"
+      LogGroupName: !Sub "/aws/amazon-msf/${ApplicationName}"
+
   FlinkApplicationLogStream:
     Type: AWS::Logs::LogStream
     Properties:
       LogGroupName: !Ref FlinkApplicationLogGroup
-      LogStreamName: "kinesis-analytics-log-stream"
+      LogStreamName: "amazon-msf-log-stream"
+
   FlinkApplicationLoggingOption:
     Type: AWS::KinesisAnalyticsV2::ApplicationCloudWatchLoggingOption
     Properties:
@@ -111,7 +103,7 @@ Resources:
               - log-group
               - Ref: FlinkApplicationLogGroup
               - log-stream
-              - kinesis-analytics-log-stream
+              - amazon-msf-log-stream
 
   MSKIAMManagedPolicy:
     Type: 'AWS::IAM::ManagedPolicy'
@@ -126,22 +118,27 @@ Resources:
               - kafka-cluster:Connect
               - kafka-cluster:AlterCluster
               - kafka-cluster:DescribeCluster
-            Resource: !Sub "arn:aws:kafka:${AWS::Region}:${AWS::AccountId}:cluster/${Clustername}/*"
+            Resource:
+              - !Sub "arn:aws:kafka:${AWS::Region}:${AWS::AccountId}:cluster/*"
           - Effect: Allow
             Action:
               - kafka-cluster:*Topic*
               - kafka-cluster:WriteData
               - kafka-cluster:ReadData
-            Resource: !Sub "arn:aws:kafka:${AWS::Region}:${AWS::AccountId}:topic/${Clustername}/*"
+            Resource:
+              - !Sub "arn:aws:kafka:${AWS::Region}:${AWS::AccountId}:topic/*"
           - Effect: Allow
             Action:
               - kafka-cluster:AlterGroup
               - kafka-cluster:AlterGroup
-            Resource: !Sub "arn:aws:kafka:${AWS::Region}:${AWS::AccountId}:group/${Clustername}/*"
+            Resource:
+              - !Sub "arn:aws:kafka:${AWS::Region}:${AWS::AccountId}:group/*"
           - Effect: Allow
             Action:
               - kafka-cluster:*TransactionalId*
-            Resource: !Sub "arn:aws:kafka:${AWS::Region}:${AWS::AccountId}:transactional-id/${Clustername}/*"
+            Resource:
+              - !Sub "arn:aws:kafka:${AWS::Region}:${AWS::AccountId}:transactional-id/*"
+
   FlinkApplicationServiceExecutionRole:
     Type: AWS::IAM::Role
     Properties:
