+++ java/Windowing/pom.xml
@@ -19,7 +19,7 @@
         <scala.binary.version>2.12</scala.binary.version>
         <kda.runtime.version>1.2.0</kda.runtime.version>
         <log4j.version>2.17.1</log4j.version>
-        <shade.mainClass>com.amazonaws.services.msf.windowing.TumblingWindowStreamingJob</shade.mainClass>
+        <shade.mainClass>com.amazonaws.services.msf.windowing.kinesis.TumblingWindowStreamingJob</shade.mainClass>
     </properties>
 
     <dependencyManagement>
